Add mechanism for selecting JSON implementation to work with.
This is a little hacky and might be removed in the future.
By setting JSON_IMPL to one of 'json', 'simplejson' or 'auto' you
can change the implementation on which utils.json will be based on.

This needs much more testing on python2.5 and python2.7, I suspect
some bugs are still there to find with this new switch.