Change Bundle addressing to use SHA1 of the content instead of the primary key.
Change DjangoAPI.get() and DjangoAPI.bundles() not to depend the primary key of
Bundle table. Instead the means of addressing bundles is by using SHA1 of the
content. This makes test results stable and generally feels better. You can
share bundle SHA1 and it will work on any database that has this document.