Improve DashboardAPI.put()
There is a comprehensive docstring that explains all the error conditions of the
function and its precise behaviour. A small API change has been made to make
put() return the SHA1 of the content instead of the database-dependent primary
key. In addition the actual bundle content is stored with a filename that is
derived from the primary key instead of the provided content_filename. This
change will simplify storage management and gets rid of the requirement to
validate the user-provided data that might have been used to derive a file name
on the file system (thus improves security)

All of the above is thoroughly tested with unit tests.