Improve XML-RPC API help page.
The sidebar contains a list of functions that link to anchors in the page.  The
page has an example python code that can interact with the server.  The view
exposes dashboard URL (the full URL of the website) using 'sites' application.
Administrators need to configure correct site for the values to make sense.