Simplify openid deployment and offer both options by default.
I changed the base.html template to offer both authentication methods. After
thinking about it for some time I think this way is better (existing users have
accounts, admin has a model account). I also changed bundle_stream_list.html
not to offer a sign-in link at all (the user probably would not have noticed it
at all as the header is same on every page so it's easier to remember that
users can sign in from there). I also removed the context processors with
LOGIN_URL as it was no longer required.

I also fixed LOGIN_REDIRECT_URL to point back to /dashboard (in default
settings) as this corresponds with the default apache configuration file.