Add load quirks to deal with 1.0 format issues out in the wild
Two format issues that are specified differently in the spec and
implementations are:
 * TestRun software context is encoded as 'sw_context' while the spec
   mandates 'software_context'
 * TestRun hardware context is encoded as 'hw_context' while the spec
   mandates 'hardware_context'

The new code adds a _load_ quirk (that can be disabled optionally) so that
those old bundles are silently fixed. The quirk is only active for bundles
that indicate the format is 'Dashboard Bundle Format 1.0'. In the future
abrek and other tools should generated data with format '1.0.1' to indicate
they expect full spec compliance.