Rewrite test loader to work better with django internal test runner.
It's now possible to run tests selectively by running manage.py test and providing a list of test identifiers.
An identifier uses this format: ApplicationName <DOT> TestClass [ <DOT> TestFunction ]

Note that module paths are not relevant, so it's not important if TestClass is
inside some.other.submodule, that part is not taken into account.