Use linaro_json instead of local copy of the same.
Due to backend change some error contidions in unit tests have been slightly
changed. As a positive side effect I found a typo in one of the tests (in the
JSON text used by the test) that caused the test to fail for a different
reason. The new backend includes line and column information in all syntax
errors.