Remove parts of LP658917 regression tests.
Those tests were meant to be moved to memory-to-database deserialization tests
but that layer changed significantly and all of those tests were using
internals that no longer exist.

Three tests are removed:

1) test_mem2db_TestCase__no_test_case

This tests checked that a database TestCase when deserialized from a object
with no test_case_id would end up with a None test_case reference. This test
should be rewritten and become a part of BundleFormatImporter_1_0 test case.

2) test_mem2db_TestCase_test_case_with_units

This test is similar and checks the usage of units field

3) test_mem2db_TestCase_test_case_without_units

This test is also similar and checks what is the default value of the units field.