Fix unit tests to be compatible with simplejson < 2.1.x
The fix is innoncent, instead of checking error type we just check for error
subclass.  Newer versions of simplejson have a dedicated class for decoding
errors that is a subclass of ValueError. We can just check for ValueError since
none of launch-control actually looks at any of the extended attributes.