Add TestDevice and TestDeviceClass models.
TestDevice instance will be necessary for each device addressable by other
components of the LAVA stack. In practice each device in the "test lab" that
can be used to perform tests will have an entry like that.

TestDeviceClass is used to represent equivalent devices. The actual devices can
be configured differently at any given time, the intent of this model is to
bind devices that are identical by hardware. It is currently assumed that all
test devices can be reconfigured for the purpose of the actual test. In the
future we may introduce another model that would help to represent situations
where a specific device (of a common class) has some additional modifications,
or is hooked to some specialized external measurement tools that cannot be
easily (and automatically) moved to other devices of this class. Currently this
is not supported.