Move parasitic_cone contents to the root directory.
The old directory was just a "place" I was using to drive new development and
as Paul Larson noted it's not really necessary. The code was therefore moved
to the top-level directory.

In addition due to the way django imports applications and projects I had to
move lava to a different directory not to cause import failures (you cannot
have a project called the same way as an existing module or package). I
therefore moved the directory to liblava.