Update attachment schema for 1.1 format.
Attachment schema was never finished before. There was a general desire to
support storage of arbitrary binary files. The schema was not defined though.
This change rewrites attachments from a
dictionary-of-file-names-mapping-to-arrays-of-lines to a proper array of
attachment objects. Each attachment object has three mandatory properties,
pathname of the original file, mime type of the contents and the base64-encoded
string with the actual data.