Add an ability to import all documents as latest format via evolution.
This option is disabled by default but can simplify the import process and
currently passes all tests. In the future we can consider discarding all but
the latest and greatest importer to simplify the codebase. Since all format
upgrades retain all information this operation is safe.

The only downside would be the lack of DocumentEvolution method that safely
converts from a more recent format to an older format if the operation is safe
and is not discarding anything. This would be needed if there are some legacy
programs that are still processing 1.0 documents manually and were not adapted
to use any abstraction libraries or more recent formats directly.

Currently we store the actual test bundle / document that the user is uploading
to the system but in the future we may choose to discard it and only re-crate
it on demand.