Make bundle_list view much more database efficient by using select_related().
Previously, for each bundle, there was an additional query that selected the
bundle deserialization error object. Now this is query is no longer needed and
entire page is generated with one bigger query using left outer join. This is
made possible by using select_related() that joins on the
BundleDeserializationError model. In addition a second join is made on
BundleStream this prevents us from re-fetching multiple copies of the single
bundle stream that all bundles share.

Based on my rough testing the running time of this page has dropped by x10 or
more, it was bound by the time required to process each of the queries for all
N bundles (in my case about 1500)