Improve bundle detail view.
Adding the sidebar with useful information makes bundles feel more as
first-class objects. This also exposes some hidden data such as the
serialization format used in the bundle (currently hard-coded to JSON in the
model) as well as the document format. Additionally the user is given a hint on
how to download the bundle for local storage/analysis/processing using a full
lava-dashboard-tool invocation.