Fix top menu alignment/rendering.
I tried a lot of things to make this work. Currently the main panel is a flex
container, each item is wrapped in a div box that can optionally flex to fill
the space (like the spacer between menu items and indicators). The actual text
of the menu is rendered as a table-cell to get vertical-align to work sensibly.
The innermost item is also sized with height: 80px to make the menu wide enough
and look nice. My attempts to use min-height to avoid overflow failed as this
does not seem to work with everything else.

Testing done in Firefox 7 and Chromium 15 on Ubuntu, pending testing on other
systems (where fonts will differ, among others) and mobile browsers.