Adjust the menu bar.
Another adjustment, apart from slight visual changes to make it more readable
there are important bug fixes in the menu javascript code. It should no longer
overflow the page and should behave properly on both sides of the screen.

There are debugging console.log() messages everywhere that have to be removed
but I wanted to keep them in this patch.