Add array backend for data tables
This simple backend performs all operations inside python's runtime
memory. It is suitable for working with small data sets and experimenting
with data-tables.

The actual backend is not fully finished as it is not essential. It lacks
per-column search, regexp support and multi-column sort (it is broken as
each sort operation is applied in reversed order)