Fix multi-column sorting in the ArrayBackend
Multi column sorting used to use naive sorting loop. When more than
one column is used in comparison we need to sort the columns in reverse
order with a stable sort function. Luckily python's list.sort() is already
stable and all that we needed to add is reverse iteration over all sort
directives