Remove length limits from TestCase model.
The limits are slowly biting us as little annoyances here and there. This time
we hit a limit of the TestCase.test_case_id field which was set to 100
characters. Like previously this limit was an arbitrary choice (where I was
hoping that CharField is more efficient / has less usage restrictions than
TextField). Luckily with PostgreSQL this distinction is cosmetic and both types
have exactly the same performance (CharField is simply a TextField with a
length check limit).

This patch changes _all_ the fields in TestCase to use TextField instead of
CharField (this affects test_case_id, name and units) and adds the necessary
south migration code.