simplify fastmodel deploy/boot
fastmodel's don't have to have the kernel,initrd, and device tree
specified. We can rely on the semi-hosting bootwrapper (axf file).
This makes for a little less code.

Also - the base client has a function called boot_linaro_image. We really
just need to implement what's required of the function it calls
_boot_linaro_image