A subtle change in flock and linaro-android-media-create
I've encountered an issue where fastmodel's don't run well from a root
user account. We normally run this part of the dispatcher as root so
that things like linaro-media-create work properly. The best hack I've
found so far is to run the fast model dispatcher as a non-root user,
and create an /etc/sudoers file like:

  %admin ALL=NOPASSWD: /usr/bin/linaro-android-media-create

This change just makes the sudoers entry smaller so it doesn't include
the flock stuff