add support for remote device types
This adds a new concept of devices that get run via celery rather
than running locally. This includes the main logic for the "master"
node, ie the code running on your main server. Another follow-on
change is required to the scheduler to handle the running of the
job when its on a remote node.