add ability for scheduler to write to stdout
The scheduermonitor command is currently set up to always log its
output to a file. However, when its run via a celery task, we need
to stream the output of the dispatcher it launches to stdout instead
so that celery can then stream it back to the "control" node where it
will then get logged to the correct file.