fix permissions on fastmodel images
We have a weird issue with the RTSM simulator binary. It does not like
to be run as root. We also have a weird issue with the dispatcher, it
pretty much has to run as root because of all the calls to mount/echo,etc
that it does when customizing the image we build.

The easiest workaround for this is to:

 1) make the FM images be group read-write
 2) make the simulator binary in the conf be something like:
    simulator_binary = sudo -u www-data <normal command args>

This commit takes care of the first part. The remaining work has to be
done in the config file itself