fix master file deletion logic
the bugs fixed by the previous commit were causing cache files to not
get deleted at exit because the exception handler for that path didn't
have the logic. The only way to delete files was if you got to the logic
where the master image was launched. This just simplifies things by doing
an 'atexit' so it won't get skipped