unify temp directory logic
We've hit a few bugs where we were leaking temp files created by the
dispatcher. This is somewhat due to the fact multiple places were
creating temp directories/files and relying on different mechanisms
to delete them upon exit.

This change makes things more explicit. The client (ie master,qemu) has
to create a temp directory. This is done using a common function,
LavaClient.get_www_scratch_dir. This function also registers an atexit
function so the directory will be deleted upon process exit.

We then update functions that use temp/cache directories to require
them as an input parameter so that the control is in the hands of the
caller and not luck or fate.