support for boot testing an openembedded image
This is a small change to allow boot testing an openembedded image. We
needed to ignore the dpkg-divert failure and also allow the job file to
specify an alternative set of boot commands since the OE images don't
use and initrd.

The way you can test this is by submitting a job whose boot action is:

  {
    "command": "boot_linaro_image",
      "parameters": {
        "options": ["boot_cmds=boot_cmds_oe"]
      }
  }					      },

then ensure you have the proper boot_cmds_oe conf set up for your
device-type.