Add device_version field to test job metadata
LavaClient received a new property called `device_version`, which is a stub and
must be implemented by subclasses.

Also included is the implementation of device_version for LavaQEMUClient, which
currently relies on the fact that qemu-system was installed via APT because the
output of `qemu-system-arm --version` is not machine-readable.

Implementation for fastmodel and master image clients will follow.