make fastmodel device a generic client implementation
The fastmodel client is now basically a generic shell that works with
a Target object. This makes it totally generic so that new devices built
completely on the Target API can still work with tradition actions in
the dispatcher

We also remove the old fastmodel_deploy action which was never used in
production.