add on/off function to master image device
This allows us to boot into the linaro image. It might make more
sense to do the "deploy_image" patches before this, but they are hard
to verify without boot support. So this was done first.

These functions were copied from the old master.py client. However,
the _boot_linaro_image logic was updated to support ubuntu/android in
one function rather than two functions. This is done by introducing a
new dictionary to the target device called "deployment_data". This is
used to set per-deployment options. For example, when the deploy_*
methods are implemented for master.py they will set the proper boot_cmds.