add notion of testdef to lava_test_shell action
This adds a testdef_url and timeout to the parameters this action
accepts. The testdef_url should point to a file that supports the
lp:lava-test's "declarative test" model. The timeout value is how
long in seconds the test will be given to run.

This series just gets the testdef in place. The next few patches
actually do something with this.

An example declarative test would be:
{
    "format": "Lava-Test Test Definition 1.0",
    "test_id": "passfail",
    "install": {
        "bzr-repos": ["lp:lava-test"],
        "git-repos": ["git://git.linaro.org/ubuntu/daily-prebuilt-images.git"],
        "steps": ["echo installing!"]
    },
    "run": {
        "steps": ["/bin/echo mem-stress-a7 : PASS"]
     },
     "parse": {
        "pattern": "(?P<test_case_id>.*-*)\\s+:\\s+(?P<result>(PASS|FAIL))",
        "fixupdict": {
            "PASS": "pass",
            "FAIL": "fail"
        }
    }
}