include new lava-test-shell scripts
These are the first stage for the basis of a more black box style approach
to testing in LAVA. There are two scripts:

* lava-test-runner(andrdoid/ubuntu): This script is intended to be
launched near the end of the boot process. Its job is to find tests
that have been configured to work. It does so by looking at the
${WORKFILE} for tests defined to run. Each test directory must
include a "run.sh" file which is to be run by the lava-test-shell.

The test runner also saves the information needed later on for the hardware
and software context information that goes into a LAVA bundle.

* lava-test-shell: This is super simple at the moment. Its responsible
for executing the actual command. It streams stdout/stderr to both the
serial console and to a file. Currently stderr is streamed to stdout.log
and this should be fixed, but its a start. We also save the exit code
for the process.