allow action to handle multiple testdef urls
The lava_test_shell code had this logic already, but the action didn't.
This changes the action to accept an array of URLs. This implied some
fixes to badness in the code:

* making the /lava/tests/<testname> handle repeats by now using an
index so its /lava/tests/01_testname, etc

* configure_target decides the name of the test install directory,
explains how the target's path and the hosts path can differ, and
then passes this path to its functions.