regression with lava-test for fastmodel
I'm not sure how this ever worked, but the recent juggling around
of prompts has indicated an issue with the default timeout set
for fastmodels. Any command that passes -1 as a timeout uses the
default which was 90, and way too small to do something like run
the bigLITTLE suite. This gets it to a sane value used by all
other device types.