have lava-test-shell try and clean up after itself
We want to try and remove our test-runner.conf file so that
the job won't get executed on the next boot. This is needed in
case a job file includes a lava_test action after the lava_test_shell
action.

In order to do this, the test-runner.conf can't be in /etc for android
because this partition is mounted read-only. So everything has been
moved to a read-write partition except for the etc stuff needed to
launch our test-runner at startup.