fix handling of bundle gathering logic
This started off with the simple goal of not letting the failure of
adding one bundle keep us from trying to add the other bundles. It
then turned into a small attempt to clean up how we deal with errors
while processing bundles.

An example of how this used to look:

 http://validation.linaro.org/lava-server/scheduler/job/27529/log_file