shorten android tester prompt
We recently made some changes to the tester prompt. We are seeing
a common failure such as:

  http://validation.linaro.org/lava-server/scheduler/job/37007/log_file

Where it seems our new, longer, prompt is often getting other logging
from the kernel jumbled into it. This shortens the prompt in hopes
of getting health jobs completing like the good-old days.