improve user experience for logged in/out access of resources
Our 403 handling stinks. You get a message saying you don't have access,
but then when you log in, it takes you to our home page rather than to
the resource you wanted.

This approach checks in the 403 logic if a user may just need to log in.
It also redirects the user back to the original resource after a good
login.