make boot options handling more flexible
boot options are currently done in a pretty rigid way that don't allow
for flexible value checking and requires code changes for updates.

This change allows them to come in a more flexible manner via the
device/device-type config file. It also includes changes to our
fastmodel code as an example.