simplify the power_off logic for a target device
The power_off logic was getting messy because its hard for the
code to know how/when it can call sync. The previous commit allows
us to not worry about the sync anymore.

In addition, it was noticed that lava-test for fastmodel's could
loose the file, so we call a sync there to be safe.