only accept jobs for boards configured in instance
This is the key piece of logic to some new changes we are adding
for better support of "remote workers". The idea is to have a type
of master/slave logic in LAVA. The main "control" server runs the
django app and can accept requests. You then can have multiple worker
nodes. Each worker node only run jobs for devices configured in its
instance using the dispatcher's "lava devices" API.

Also in this commit is the removal of using celery. Now that we are
running multiple scheduler instances, celery isn't needed for dispatching
job requests. Instead each worker node will now have the "lava-logs"
directory of the master node mounted over sshfs so that jobs are
run remotely but logs are kept centrally.