improve some android boot logic and options
not all android devices want the coming up of the home screen to
be counted as a good boot. So this adds a config option to make
waiting for the home screen optional. Additionally, I noticed we wait
for the homescreen when establishing a tester session which makes no
sense since you've already done that waiting.

Lastly, I made the number of loops to wait for home screen configurable
in case someone gets the urge to increase/decrease for their device