introduce global delay in our serial console send routine
We've seen issues where sending things like 16 byte or bigger strings
can often get messed up in serial console implementations. This has
been seen with our PS1 prompt on panda and seems to be fairly consistent
on TC2 in a couple of places.

Introducing a small .05 second delay seems to make things work well
on TC2 and is small enough to not really annoy anyone. So this seems
worth fixing in general.

I also fixed a few pylint warnings in this class while I was looking
at it.