Now actually capture all the output in both stdout and log file.
The problem was that subprocess.call and subprocess.check_call bypass the
write() method of the file objects you pass in stdout and stderr. Instead the
subprocess functions will use the underlying file descriptor (#fileno()) and
dump the output there, so you need a file object that has a proper descriptor
to duplicate the output to both stdout and the output.txt log file.

This can be accomplished by creataing a pipe to tee(1); that pipe gets a proper
file descriptor and will spit the output both to stdout and to the file you
pass to it in the command line.