force finishing of emulated devices
With both fastmodels and QEMU we have the risk of leaving runaway processes
after a job finishes, what may block resources and impact future usage of the
same (virtual) device (and it just wrong). This commit generalizes the solution
I found before for QEMU and applies it to fastmodels as well.