device/target: abort boot without deployment
Whether we are booting through boot_linaro_image or through
lava_test_shell, the common denominator is that the action will need to
access the deployment_data. The fact that there is no deployment data
means no deployment was done, so we abort immediately instead of waiting
for errors when trying to read data from an empty deployment data
dictionary.

This fixes bug #1224262

Change-Id: I479f426e1cca7df9393f0103a798f595fe0b8ef8