Add ARM Base and Cortex model support.
New models added:

* rtsm_fvp_base-aemv8a
  - 8 x A57
  - Bridged NIC
* rtsm_fvp_base_cortex-a57x4-a53x4
  - 4 x A57
  - 4 X A53
  - Bridged NIC
* rtsm_fvp_base_cortex-a57x1-a53x1
  - 1 x A57
  - 1 x A53
  - Bridge NIC

Notes for this change set:

* The base and cortex models chain load secure firmware. Therefore the model must be launched in
same directory as _ALL_ the firmware. I have asked ARM and you cannot explicitly pass the path to
each piece of firmware.

* Switch the LMC arg --dev vexpress -> --dev fastmodel. The fastmodel LMC device target now has
proper support upstream. (2013.09)

* Use a common binary naming convention across all fast models.
  - Kernel -> Image
  - Ramdisk (Optional) -> Initrd
  - The binaries are renamed on the sd image as well as the output directory for consistency.

The problem is that a ramdisk for instance might be named initrd-3.12-rc1-234490.img on the image,
making it impossible to have static boot_cmds to reference it. By renaming this binary using this
common naming convention we can have static universal boot_cmds to work in any case. It also helps
when creating new boot commands, as you don't have to worry about how the binaries are named on
the image / output directory.

* Per Fathi's recommendation switch LAMC device arg from vexpress-a9 -> vexpress.

Change-Id: Ieb02f8b962f7582b5b7de24d0f97619b87479d98
Signed-off-by: Tyler Baker <tyler.baker@linaro.org>