Master/Slave scheduler implementation.
The master scheduler identified by the absence of worker.conf file
will assign devices and schedules jobs in the database.

The slave scheduler identified by the presence of worker.conf file
will just read from the database and looks for jobs that could be run
by the devices attached to it.

This is a multinode job agnostic greedy scheduler. Devices are
assigned with a greedy approach, but the difference is that, a multinode
group is taken as a whole and assigned as much devices as possible
before moving to the next job. This is better than checking for 100%
device availability before assigning devices to multinode jobs, since
waiting for 100% availability leads to starvation in an environment
where both multinode and single node jobs co-exists (tested with a
patch). OTOH, the new approach works just fine since we have the master
scheduler which is the only one, scheduling devices to jobs, unlike the
previous scheduler where every scheduler was messing up with the
database.

Change-Id: Ie885b82d64ca5b74cf0812aeaa34dae8c625adf5