Improved prompt handling.
wait_for_prompt updated to handle testing for false positives.
  New device configuration parameters:
    bootloader_prompt_is_re (bool. Default = false. Optional)
    bootloader_prompt_test_send (text to send when we think we have a
      prompt. Optional).
    bootloader_prompt_test_response (text to expect having sent the
      text in bootloader_prompt_test_send. Optional)
Added the ability to test for a return code if it isn't in the prompt.
  New device configuration parameters:
    tester_rc_cmd (string)
Allow customisation of device deployment data in the device config.
  New device configuration parameters, all of which overwrite the
  device deployment data of the same name if set:
    tester_ps1_includes_rc
    tester_ps1
    tester_ps1_pattern

Allow modification of deployment data (possibly change this later so the object is immutable but you can provide values to the constructor to overwrite defaults).

Removed unused delay variable.

Change-Id: Iaeef1917b9d1615784e50dcd2c15edcd2a4a79a6