When taking a board offline, notify when current job completes
Bug #1224939. Adds an admin_notification field to TestJob which
is only populated if a user with admin privileges puts the device
running the job into maintenance mode. The email address of that
user is then used to send a completion notice when the job finishes,
whether the job was incomplete or not. Any notification list in the
JSON is independent and only one admin notification is allowed per
transition.

Change-Id: I3daba03b6b09257126876b5de8bf07c2b7acaec8
Reviewed-on: https://staging.review.linaro.org/299
Reviewed-by: Antonio Terceiro <antonio.terceiro@linaro.org>