Check filesystem is writeable rather than partition
Writing data directly to a partition destroys the filesystem there and causes the /dev/disk/by-label symlinks to be lost.
Instead, mount the filesystem and try to read/write a file to it using dd with direct iflag/oflag and then md5sum what comes back.
We have to md5sum the data as it could be garbage and then grep for the right md5sum. Previously grep was trying to match binary data - not good.

This needs merging pretty fast as it's broken quite a lot of the filesystem labels on staging devices. Argh.

Change-Id: Idfc01fa1720eee09a1d8a6d4e22811c119feea7e
Reviewed-on: https://staging.review.linaro.org/341
Reviewed-by: Antonio Terceiro <antonio.terceiro@linaro.org>