Expand and document support for delayed serial comms.
In cases where the serial controller is unable to cope
with bursts of serial input, provide a pair of device_type
configuration options to separate each character sent
over the serial line by a delay in milliseconds.

Rename the existing PXE support to clarify that this is bootloader
specific and implement as new support for delays within the test
shell.

Use name of bootloader_serial_delay_ms for bootloader parameter and
test_shell_serial_delay_ms for the test shell parameter.

Add the test shell delay to the default highbank device_type
configuration. Fixes bug #1229644

Add documentation on both options.

Change-Id: Ia9b54818d12625a5e8f70d023bbc8428be662de7
Reviewed-on: https://staging.review.linaro.org/359
Reviewed-by: Tyler Baker <tyler.baker@linaro.org>