Remove non-POSIX shell calls in multinode calls
Use sed to replace non-POSIX use of ## and %%.
Allow for Android not having a directory for the
cache file.
Use awk to replace grep -P which requires libpcre.
Replace _get_key_value_pattern to use grep.
Ensure lava-group exits zero when not in error
and handle the filter option correctly.

Fixes #1242692 - MultiNode on Android.

Change-Id: I9a49a803e8da36b9db900a291cd35510e107e526
Reviewed-on: https://staging.review.linaro.org/368
Reviewed-by: Antonio Terceiro <antonio.terceiro@linaro.org>