Update packaging for LMP support files
Add missing lava-network script to the packaging.
Set the path to the LMP support files at the same
  time as the path to the MultiNode support files.
Ensure LMP and MultiNode support files exist for packaging.
Final changes for LAVA-97 in lava-dispatcher.
Use a review-specific branch.
Use the same location for packaging as with deployments
  for the lava_test_shell support files to avoid
  hardcoding paths.
Fix problem with data_files not supporting directories,
  package all files in lava_test_shell but not dirs.

Change-Id: I75d89aad95ca04d3262d62c5ca3c7c8ff8342f0f