kill adb process before retry when there is an adb connection problem
There is a case that we will have "error: protocol fault (no status)"
error with adb command, and to resolve it we need to kill the adb process.

so here when we caught the ADBConnectError exception, before we try to
boot the android image again, we will first check if there is an adb
command problem with the adb command via "adb devices" command,
if so we will kill it first and then try to boot the android image again.

Change-Id: I86e635cd006de1cd1c4f25dd1dbcc55a5b52296b
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Reviewed-on: http://staging.review.linaro.org/508
Reviewed-by: Vishal Bhoj <vishal.bhoj@linaro.org>
Reviewed-by: Dave Pigott <dave.pigott@linaro.org>
Reviewed-by: lava-bot <lava-bot@linaro.org>
Reviewed-by: Antonio Terceiro <antonio.terceiro@linaro.org>