LAVA-832 - Allow adding comments to the device status transition log.
An user who has access to administer a particular device will have access to
edit the device state transition messages.

This patch introduces in place editing on a text! wherein a user has to click
the device state transition message text in order to edit (explained in
tool-tip). Saving happens when the user presses enter, without refreshing the
entire page!!

Change-Id: I4b68eb12aa449ad8f938db647f50333c7bcc340d
Reviewed-on: http://staging.review.linaro.org/606
Reviewed-by: lava-bot <lava-bot@linaro.org>
Reviewed-by: Neil Williams <neil.williams@linaro.org>