Periodic kills of canceling jobs
Write out the PID of the lava-dispatch call to the job output directory
Check for jobs with status CANCELING during each check of the JobList
If pidrecord file exists, send SIGTERM to that process group and unlink.
Call TestJob.cancel() to put the database into the correct state.
Cleanup the pidrecord file at exit
Obtain the pgid and pass this to killpg to kill all child processes
of the schedulermonitor process for the job.
Possible fix for database connection failure problems.

Change-Id: I0b57c14ba84fdb2591e2bf3322ed5b9f37357e63
Reviewed-on: https://review.linaro.org/750
Reviewed-by: lava-bot <lava-bot@linaro.org>
Reviewed-by: Tyler Baker <tyler.baker@linaro.org>