Ensure multinode targets are allowed devices
Closes: #LP1266614 - multinode role errors
allowed_devices needs to track the device_type and
then this value can be used to pick the first allowed
device of a specified type as the target of the TestJob
created for the matching role.
Ensure requested_device is None so that devices are not
reserved whilst already running a job.

Fix typo in _process_multinode which used the wrong
job details to choose the device for the requested
device type.

Change-Id: I96376e43dbe2d6f073bd267e05bb1ce2e10ca8ed
Reviewed-on: https://review.linaro.org/839
Reviewed-by: lava-bot <lava-bot@linaro.org>
Reviewed-by: Tyler Baker <tyler.baker@linaro.org>