lava_command_run: fix logic for failed commands
A failed command was causing an exception by trying to use the `rc`
variable which was never assigner because the body of the try block
raised an exception. since session.run already does all the checking we
need, we can handle any exception it throws and log it as an error,
going on with the rest of the commands.

Change-Id: Ia16f8d7c886a2f0e5dbbef607607a3b74db23b39
Reviewed-on: https://review.linaro.org/858
Reviewed-by: Tyler Baker <tyler.baker@linaro.org>