Clear deadlocks in multinode scheduling.
This is an add-on to the greedy scheduling approach which is prevailent in LAVA
scheduler as of today. There is a point when the greedy scheduler approach will
create a deadlock, which cannot be avoided. But this patch helps to recover
from the deadlock after 5 minutes of the deadlock occurance.

NOTE: As mentioned in (d8542310733f519e91683c9f4742444b14350f83), greedy
approach is the way to go, but this is one exceptional deadlock case whose
probability of occurance is very rare!

Change-Id: Ibc06754f241dd345b5130a6f6a0d38223e935ec2