bootloader.py: u-boot: ensure ramdisk has proper u-boot header
If passing in a raw ramdisk, u-boot will not be able to properly pass
the ramdisk using bootm/bootz unless it has a proper u-boot header. To
fix, check the ramdisk to see if it has a u-boot header, if not add
one using the 'mkimage' tool.

NOTE: this adds a requirement on the 'mkimage' tool which is
      distributed as part of the u-boot-tools package.

Change-Id: I884c5d9839f7ee6d6f863905f8a3b178eb5266e0
Signed-off-by: Kevin Hilman <khilman@linaro.org>