Metadata changes for Image Reports 2.0
Add the ability to add which custom attributes will be 'tracked' in the Image report so when some meta data changes from build to build, it is displayed in the report so the user can see at which point there was a change.
Add a visible marker for comments in the report with a tooltip containing the comment.

Change-Id: Ia7d8361287b8635fd38753c426b0b46d329fdc31