Retry transaction commit for job completion.
This is required since worker heartbeat updates are frequent and sometimes
may interrupt with other saving job and device status during job completion.
Though this is observed in certain versions of django, it is safe to have
it here in order to avoid any potential deadlocks in vulnerable django
versions.

Change-Id: I38b8bf8e29c10911686f5b2ab351163cfea7bce6