dispatcher: add vm-groups support
Add support for passing data from the vm_group host, through the
existing MultiNode API and allow collection of that data by all vm_group
guests to allow each guest to start running the LavaTestJob once they
have the IP address of the host.

A message is emitted from within deploy_linaro_image on the host device,
and caught by the waiting NodeDispatcher for each guest in the same
vm_group. The messsage contains the IP address of the host and signifies
that the host is ready for the guest to connect and start the VM.

Signed-off-by: Neil Williams <neil.williams@linaro.org>
Signed-off-by: Antonio Terceiro <antonio.terceiro@linaro.org>

Change-Id: Ib643b687f13d2c44e18722b150d56cb29c9dabf7
(cherry picked from commit 568c96258cdea03dc69e37694331186f9cdac355)