Add LAVA_NETWORK_IFACE to override LAVA_SERVER_IP
With a dispatcher based on a laptop, it is common to
need to change the LAVA_SERVER_IP each time the network
lease updates. Instead, each time lava-dispatch starts,
the current IP address of the specified interface can
be retrieved by lava-dispatch itself. If no value is
specified, no action is taken and LAVA_SERVER_IP is
used directly but LAVA_NETWORK_IFACE could become
the default way to configure the dispatcher network options.

LAVA_NETWORK_IFACE supports a list of possible interfaces.

Use the LAVA_SERVER_IP to update LAVA_IMAGE_URL so
that the downloads use the correct location.
Allow an error in this setting to be written to the Job
logfile which will be visible in the Complete Log.

Change-Id: Ic0e0cd0b1e437c2ef147e13eb2def912e56e6084